Move spec class to utils directory
